--- src/routes/App.tsx
@@ -1,10 +1,0 @@
-import '@fontsource/abril-fatface'
-import '@fontsource-variable/outfit'
-
-import { ChannelPage } from './ChannelPage'
-
-function App() {
-  return <ChannelPage />
-}
-
-export default App
